Reorganize directory structure and refactor
