+++ uapi/base_uapi/utils/misc.py
